split files in landscape plugin
